Common: add fwdtrackUtilities.h
